Work on reporting (#17)
* Address some TODOs

* Rename report and add some methods text

* Add results table to report

* Explore model specs

There is weak evidence that planned synthesis predicts treatment assignment, but reasonable evidence that field does. So, temporarily removing planned synthesis from the model of endogenous treatment assignment.

The analysis for rec_vs_nonrec for resource use seems to dramatically overestimate the effect size if we perform the planned analysis. So, exploring using interval regression and not accounting for endogenicity for that outcome.

These decisions are temporary because I'm waiting to hear from Jose about whether data on the existence of protocols is available and is likely to be useful.

* Add figures to report

* Add table note

* Add full regression results to report